+++ example/lib/src/test_pagination.dart
@@ -1,4 +1,4 @@
-import 'dart:developer';
+import 'dart:async';
 
 import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
@@ -6,10 +6,13 @@ import 'package:sura_flutter/sura_flutter.dart';
 import 'package:sura_manager/sura_manager.dart';
 
 class SuraManagerWithPagination extends StatefulWidget {
-  const SuraManagerWithPagination({Key? key}) : super(key: key);
+  final FutureManager<int> dataManager;
+  const SuraManagerWithPagination({Key? key, required this.dataManager})
+      : super(key: key);
 
   @override
-  _SuraManagerWithPaginationState createState() => _SuraManagerWithPaginationState();
+  _SuraManagerWithPaginationState createState() =>
+      _SuraManagerWithPaginationState();
 }
 
 class _SuraManagerWithPaginationState extends State<SuraManagerWithPagination> {
@@ -56,8 +59,8 @@ class _SuraManagerWithPaginationState extends State<SuraManagerWithPagination> {
   @override
   void initState() {
     fetchData();
-    userManager.addListener(() {
-      log(userManager.toString());
+    widget.dataManager.addListener(() {
+      infoLog("Data manager:", context);
     });
     super.initState();
   }
@@ -113,11 +116,17 @@ class UserResponse {
 
   UserResponse({this.pagination, required this.users});
 
-  bool get hasMoreData => pagination != null ? users.length < pagination!.totalItems : false;
+  bool get hasMoreData =>
+      pagination != null ? users.length < pagination!.totalItems : false;
 
   factory UserResponse.fromJson(Map<String, dynamic> json) => UserResponse(
-        users: json["data"] == null ? [] : List<UserModel>.from(json["data"].map((x) => UserModel.fromJson(x))),
-        pagination: json["pagination"] == null ? null : Pagination.fromJson(json["pagination"]),
+        users: json["data"] == null
+            ? []
+            : List<UserModel>.from(
+                json["data"].map((x) => UserModel.fromJson(x))),
+        pagination: json["pagination"] == null
+            ? null
+            : Pagination.fromJson(json["pagination"]),
       );
 }