+++ lib/src/sura_manager_provider.dart
@@ -1,9 +1,5 @@
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:sura_manager/src/type.dart';
-
-typedef OnManagerError = Function(FutureManagerError, BuildContext);
-typedef ManagerErrorBuilder = Function(FutureManagerError, AsyncCallback?);
 
 class SuraManagerProvider extends InheritedWidget {
   const SuraManagerProvider({
@@ -20,7 +16,7 @@ class SuraManagerProvider extends InheritedWidget {
   ///Error widget use in [Manager] class
   final ManagerErrorBuilder? errorBuilder;
 
-  //A callback function that run if FutureManagerBuilder has an error
+  ///A callback function that run if FutureManagerBuilder has an error
   final OnManagerError? onFutureManagerError;
 
   static SuraManagerProvider? of(BuildContext context) {